Use custom dataclasses for ingredients and recipes
