+++ src/recipebook.py
@@ -1,14 +1,16 @@
+from dishdeck_dataclasses import *
+
 class RecipeBook:
 	def __init__(self):
 		
-		# This is a dictionary of dictionaries.
-		# Each recipe name points to a dictionary containing:
-		# source, description, ingredients, and instructions
-		self.__recipes = {}
+		# This is a dictionary of Recipes.
+		# Each recipe name points to a Recipe object containing:
+		# name, source, description, ingredients, and instructions
+		self.__recipes: dict[Recipe] = {}
 
 	def import_recipe(self, filepath: str):
-		recipe_name = ""
 		recipe_data = {
+			"name": "",
 			"source": "",
 			"description": "",
 			"ingredients": [], # Each element should be in the form: [name, quantity]
@@ -17,7 +19,7 @@ class RecipeBook:
 		with open(filepath, 'r') as file:
 			lines = file.readlines()
 
-			recipe_name = lines[0].strip()[2:]
+			recipe_data["name"] = lines[0].strip()[2:]
 			recipe_source = lines[1].strip()
 			if recipe_source != "":
 				# Trim off the markdown link label and parenthesis
@@ -36,13 +38,17 @@ class RecipeBook:
 					continue
 				if key == "Ingredients":
 					current_line = current_line[6:] # Trim checkbox and spaces
-					recipe_data["ingredients"].append(current_line.split(" | "))
+					ingredient_data = current_line.split(' | ') # Separate ingredient name and qty/unit
+					ingredient_data = [ingredient_data[0]] + ingredient_data[1].split() # Separate qty and unit
+					print(ingredient_data)
+					recipe_data["ingredients"].append(Ingredient(*tuple(ingredient_data)))
 					continue
 				if key == "Instructions":
 					current_line = current_line[3:] # Trim instruction number
 					recipe_data["instructions"].append(current_line)
 		
-		self.__recipes[recipe_name.lower()] = recipe_data
+		new_recipe = Recipe(recipe_data["name"], None, recipe_data["description"], recipe_data["ingredients"], recipe_data["instructions"])
+		self.__recipes[recipe_data["name"].lower()] = new_recipe
 
 	# TODO: Add filtering, i.e. make parameters ingredients and strict functional
 	#		strict=True means ONLY the listed ingredients can be in returned recipes, no others
@@ -52,9 +58,9 @@ class RecipeBook:
 	def get_recipe_desc(self, recipe_name: str):
 		# TODO: replace with try/catch for error handling
 		if recipe_name in self.get_recipe_names():
-			return self.__recipes[recipe_name]["description"]
+			return self.__recipes[recipe_name].description
 		
 	def get_recipe_ingredients(self, recipe_name: str):
 		# TODO: replace with try/catch for error handling
 		if recipe_name in self.get_recipe_names():
-			return self.__recipes[recipe_name]["ingredients"]
+			return self.__recipes[recipe_name].ingredients
